% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cli}
\alias{cli}
\title{cli data}
\format{
A matrix:
\describe{
...
}
}
\source{
\url{NULL}
}
\description{
The cluster information of ccRCC sample from TCGA, Obradovic et al.’s cohort and E-MTAB-1980
}
\details{
Success!!
}
