% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictcluster.R
\name{predictcluster}
\alias{predictcluster}
\title{Title predictcluster}
\usage{
predictcluster(ccRCCdata)
}
\arguments{
\item{ccRCCdata}{ccRCCdata is a gene expression matrix form rna-seq or array. rowname is gene name and colname is sample's name. The value need log2(data+1) process before predicted.}
}
\description{
Title predictcluster
}
\examples{
cluster <- predictcluster(test)
}
