% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test}
\alias{test}
\title{test data from Obradovic et al.’s cohort for test}
\format{
A matrix:
\describe{
...
}
}
\source{
\url{NULL}
}
\description{
a rna-seq data from Obradovic et al.’s cohort for test
}
\details{
Success!!
}
